backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
pymupdf
faiss-cpu
llama-cpp-python
numpy
FlagEmbedding
